검색 부분 및 시뮬레이션 검색 처리 중

+++ app/src/main/java/com/quantec/moneypot/activity/Main/Fragment/Tab3/aaa/Code.java
@@ -17,7 +17,7 @@ public class Code {
     private Integer type;
     @SerializedName("weight")
     @Expose
-    private Integer weight;
+    private float weight;
 
     public String getCode() {
         return code;
@@ -43,12 +43,11 @@ public class Code {
         this.type = type;
     }
 
-    public Integer getWeight() {
+    public float getWeight() {
         return weight;
     }
 
-    public void setWeight(Integer weight) {
+    public void setWeight(float weight) {
         this.weight = weight;
     }
-
 }
